requirements.txt
bs4==0.0.1
requests==2.12.1
pyDes==2.0.1
python-telegram-bot==9.0.0

